requirements.txt
sphinxcontrib-issuetracker
click
requests
pytest
pytest-cov
pytest-recording
codecov
vcrpy
